chore: Split command/tool.rs into tool/dispatch.rs + tool/registry.rs [ORB-10389]
Agent-Model: claude-opus-5
