+++ crates/orbit-core/src/command/tool/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod dispatch;
